requirements.txt
pandas==2.1.0
numpy==1.24.3
scikit-learn==1.3.0
transformers==4.33.0
torch==2.0.1
streamlit==1.27.0
plotly==5.17.0
sdv==0.19.0
fhir.resources==6.2.0
requests==2.31.0
python-dotenv==1.0.0
